+++ subpackages/cli/source/odood/cli/commands/assembly.d
@@ -4,14 +4,17 @@ private import std.json;
 private import std.exception: enforce;
 private import std.stdio: writefln, writeln;
 private import std.array: empty;
+private import std.format: format;
 
 private import colored;
 private import commandr: Option, Flag, ProgramArgs;
 
 private import odood.lib.assembly: Assembly;
 private import odood.lib.project: Project;
+private import odood.utils.addons.addon: OdooAddon;
 private import odood.git: parseGitURL;
 private import odood.cli.core: OdoodCommand, OdoodCLIException;
+private import odood.cli.utils: printLogRecordSimplified;
 
 
 class CommandAssemblyInit: OdoodCommand {
@@ -94,6 +97,7 @@ class CommandAssemblyLink: OdoodCommand {
     }
 }
 
+
 class CommandAssemblyPull: OdoodCommand {
     this() {
         super("pull", "Pull updates for this assembly.");
@@ -117,6 +121,68 @@ class CommandAssemblyPull: OdoodCommand {
 }
 
 
+class CommandAssemblyUpgrade: OdoodCommand {
+    this() {
+        super("upgrade", "Upgrade assembly (optionally do backup, pull changes, update addons).");
+        this.add(new Flag(
+            null, "backup",
+            "Do backup of all databases"));
+        this.add(new Flag(
+            null, "skip-errors", "Continue upgrade next databases if upgrade of db had error."));
+    }
+
+    public override void execute(ProgramArgs args) {
+        auto project = Project.loadProject;
+        enforce!OdoodCLIException(
+            !project.assembly.isNull,
+            "Assembly not initialized!");
+        auto assembly = project.assembly.get;
+
+        if (args.flag("backup"))
+            foreach(db; project.databases.list)
+                project.databases.backup(db);
+
+        assembly.pull;
+        assembly.link();
+
+        auto start_again=false;
+        if (project.server.isRunning) {
+            project.server.stop;
+            start_again=true;
+        }
+
+        bool error = false;
+        OdooAddon[] addons = project.addons.scan(assembly.dist_dir, recursive: false);
+        foreach(db; project.databases.list) {
+            auto error_info = project.server.catchOdooErrors(() {
+                project.lodoo.addonsUpdateList(
+                    dbname: db,
+                    ignore_error: true
+                );
+                project.addons.update(db, addons);
+            });
+
+            if (error_info.has_error) {
+                error = true;
+                writeln("Following errors detected during assembly addons update for database %s:".format(db.yellow).red);
+                foreach(log_line; error_info.log)
+                    printLogRecordSimplified(log_line);
+
+                if (!args.flag("skip-errors"))
+                    throw new OdoodCLIException(
+                        "Assembly upgrade for database %s failed!!".format(db));
+            }
+        }
+
+        if (start_again)
+            project.server.start;
+
+        if (error)
+            throw new OdoodCLIException("Assembly upgrade failed!");
+    }
+}
+
+
 class CommandAssembly: OdoodCommand {
     this() {
         super("assembly", "Manage assembly of this project");
@@ -125,6 +191,7 @@ class CommandAssembly: OdoodCommand {
         this.add(new CommandAssemblySync());
         this.add(new CommandAssemblyLink());
         this.add(new CommandAssemblyPull());
+        this.add(new CommandAssemblyUpgrade());
     }
 }